requirements.txt
langchain==0.2.10

langchain-openai==0.1.17 # third party package which intergrate langchain to the llm provided by OpenAI
langchain_ollama==0.1.0

langchain-community==0.2.9 # community contributed code
langchainhub==0.1.20 # an extension of langchain to download dynamically prompts that are contributed by the community
black==24.4.2 # code formatter
python-dotenv==1.0.1 # load environment variables from .env file
requests